Fix ring volume issue by addressing focus regression in keyguard.
This fixes a regression where ring volume can be changed in keyguard.
Because KeyguardHostView is now being re-created in onScreenTurnedOff(),
it loses focus and the volume keys get handled by the fallback handler.

The fix is to ensure at least one child under KeyguardHostView has focus
whenever we re-create it.

Fixes bug 7546960

Change-Id: I34b7db402401a824f463d35d7546c05dc2979243

Conflicts:
	policy/src/com/android/internal/policy/impl/keyguard/KeyguardHostView.java